requirements.txt
# Discord Bot
py-cord==2.6.1

# LLM
groq==0.4.2
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1